🚀 alpha release 298.3 (298203)

+++ src/com/discord/media/engine/S.java
@@ -5,7 +5,7 @@ import kotlin.jvm.functions.Function2;
 // $VF: synthetic class
 public final class S implements Function2 {
    public final MediaEngineModule d;
-   public final int e;
+   public final double e;
 
    public final Object invoke(Object var1, Object var2) {
       return <unrepresentable>.d(this.d, this.e, (String)var1, (Integer)var2);
